build(bazel): wire committed test data as bazel data deps for three go_test targets
The approval, filereview, and harness test targets read committed corpus/
fixture files that were never declared as bazel data deps, so they could
not run in the sandbox and were excluded from the ci.bazel nightly
wildcard. Wire the data through exported filegroups + `# keep` data attrs,
teach the path resolvers to use TEST_SRCDIR/TEST_WORKSPACE under bazel
(plain `go test` behavior unchanged), consolidate the approval package's
five duplicated corpus resolvers into one helper, and delete the
exclusion list. Also heals main-pre-existing gazelle staleness in
skill/controller left by PR #822.

Closes #722
