+++ apis/testdata/BUILD.bazel
@@ -1,0 +1,11 @@
+# Hand-written (gazelle does not manage non-Go data packages): the shared HITL
+# corpus consumed as a bazel data dep by the Go cross-edition parity suites
+# (oss#722). The TS and Java editions read the same files working-directory.
+filegroup(
+    name = "hitl_corpus",
+    srcs = glob(["hitl/**"]),
+    visibility = [
+        "//backend/services/stigmer-server/pkg/domain/agentexecution/approval:__pkg__",
+        "//backend/services/stigmer-server/pkg/domain/agentexecution/filereview:__pkg__",
+    ],
+)
