+++ backend/services/runner/fixtures/BUILD.bazel
@@ -1,0 +1,8 @@
+# Hand-written (the runner is a TS service outside bazel's build graph): the
+# golden IPC wire fixture, exposed so the Go IPC-mirror conformance test can
+# declare it as a bazel data dep (oss#722). Regenerated by `make gen-ipc-fixtures`.
+filegroup(
+    name = "ipc_protocol_fixtures",
+    srcs = ["ipc-protocol.generated.json"],
+    visibility = ["//test/integration/harness:__pkg__"],
+)
