+++ backend/services/stigmer-server/pkg/domain/agentexecution/approval/sequence_corpus_test.go
@@ -4,7 +4,6 @@ import (
 	"encoding/json"
 	"os"
 	"path/filepath"
-	"runtime"
 	"sort"
 	"strings"
 	"testing"
@@ -285,14 +284,7 @@ func parseAction(t *testing.T, s string) agentexecutionv1.ApprovalAction {
 
 func sequencesDir(t *testing.T) string {
 	t.Helper()
-	_, thisFile, _, ok := runtime.Caller(0)
-	if !ok {
-		t.Fatal("runtime.Caller failed; cannot locate the sequence corpus")
-	}
-	// thisFile: backend/services/stigmer-server/pkg/domain/agentexecution/approval/sequence_corpus_test.go
-	// repo root is seven directories up; the corpus lives under apis/testdata/hitl.
-	repoRoot := filepath.Join(filepath.Dir(thisFile), "../../../../../../..")
-	return filepath.Join(repoRoot, "apis", "testdata", "hitl", "sequences")
+	return filepath.Join(hitlCorpusDir(t), "sequences")
 }
 
 func loadSequence(t *testing.T, path string) sequenceFile {
